+++ app/src/main/java/ar/ort/edu/proyecto_final_grupo_4/domain/repository/DayOfWeekRepository.kt
@@ -4,6 +4,6 @@ import ar.ort.edu.proyecto_final_grupo_4.domain.model.DayOfWeek
 
 interface DayOfWeekRepository {
     suspend fun insertDayOfWeek(day: DayOfWeek)
-    suspend fun getDaysForSchedule(scheduleId: Int): List<DayOfWeek>
-    suspend fun deleteDaysForSchedule(scheduleId: Int)
+    suspend fun getDaysForSchedule(scheduleId: Long): List<DayOfWeek>
+    suspend fun deleteDaysForSchedule(scheduleId: Long)
 }
